Update messages.json
First, I'm unable to find where I can message the owner to ask questions that I couldn't find answers to in the ReadMe or other How To documentation sections.  Would someone please point me in that direction?   I think this is the program I've been searching for, but I'm struggling in a couple of areas.  Thank you very much.

Intention: 
  Edited for grammar, continuity, and clarity.  

Background:
  Technical Writing, Process-Flow Engineering, Proofreading, Graphical Proofing, Linguistics, Psychology, Programming (limited experience BASIC (old version), VBA (current); high interest, and can but not write from scratch; currently obtaining education in this realm), Designer, End-User Training, and much more.

Qualifications (as in information that may limit the strength of this proposal): 
  I'm a first-time user, and this is my first experiment/contribution. I have the Hello and the GitHub 101 pages open (AND bookmarked), but I've yet to read thoroughly, and am aware that proposing changes prior to being educated is less than ideal, but I got a bit enthusiastic and...well, here's my proposal.  I apologize for any issues.
 
Is there a way to test this script before proposing the change?   

While browsing for how to use with a touch screen, I came across this section, and though I'd REALLY like to test this proposal prior to submitting, I can't figure out how, and I am unable to download the Developer extension because it says something like Manifest missing.  I will look into this more.

My concerns about not being able to test this proposal are related to the quality of verbiage flow and the clarification of information presented, ensuring that the flow of phrases having subsequent phrases are not interrupted by errant punctuation or improper word forms.  

I WILL get myself educated sooner rather than later!

Thanks for considering this proposal.  If not submitted properly, please let me know.  Thank you.